+++ tests/supports.rs
@@ -18,24 +18,29 @@ pub struct Variant {
     skip: Option<bool>,
 }
 
-#[test]
-fn expansion() {
-    let di = syn::parse_str(
-        r#"
+#[derive(Debug, FromDeriveInput)]
+#[darling(attributes(from_struct), supports(struct_named))]
+pub struct StructContainer {
+    data: ast::Data<(), syn::Field>,
+}
+
+mod source {
+    use syn::{self, DeriveInput};
+
+    pub fn newtype_enum() -> DeriveInput {
+        syn::parse_str(
+            r#"
         enum Hello {
             World(bool),
             String(String),
         }
     "#,
-    ).unwrap();
+        ).unwrap()
+    }
 
-    Container::from_derive_input(&di).unwrap();
-}
-
-#[test]
-fn unsupported_shape() {
-    let di = syn::parse_str(
-        r#"
+    pub fn named_field_enum() -> DeriveInput {
+        syn::parse_str(
+            r#"
         enum Hello {
             Foo(u16),
             World {
@@ -43,7 +48,41 @@ fn unsupported_shape() {
             },
         }
     "#,
-    ).unwrap();
+        ).unwrap()
+    }
 
-    Container::from_derive_input(&di).unwrap_err();
+    pub fn named_struct() -> DeriveInput {
+        syn::parse_str(
+            r#"
+        struct Hello {
+            world: bool,
+        }
+    "#,
+        ).unwrap()
+    }
+
+    pub fn tuple_struct() -> DeriveInput {
+        syn::parse_str("struct Hello(String, bool);").unwrap()
+    }
+}
+
+#[test]
+fn enum_newtype_or_unit() {
+    // Should pass
+    Container::from_derive_input(&source::newtype_enum()).unwrap();
+
+    // Should error
+    Container::from_derive_input(&source::named_field_enum()).unwrap_err();
+    Container::from_derive_input(&source::named_struct()).unwrap_err();
+}
+
+#[test]
+fn struct_named() {
+    // Should pass
+    StructContainer::from_derive_input(&source::named_struct()).unwrap();
+
+    // Should fail
+    StructContainer::from_derive_input(&source::tuple_struct()).unwrap_err();
+    StructContainer::from_derive_input(&source::named_field_enum()).unwrap_err();
+    StructContainer::from_derive_input(&source::newtype_enum()).unwrap_err();
 }